+++ airbyte-integrations/bases/base-java/src/main/java/io/airbyte/integrations/destination_async/BufferManager.java
@@ -5,20 +5,15 @@
 package io.airbyte.integrations.destination_async;
 
 import io.airbyte.integrations.destination.buffered_stream_consumer.RecordSizeEstimator;
-import io.airbyte.integrations.destination_async.MemoryBoundedLinkedBlockingQueue.MemoryItem;
-import io.airbyte.integrations.destination_async.buffers.Batch;
+import io.airbyte.integrations.destination_async.buffers.BufferDequeue;
+import io.airbyte.integrations.destination_async.buffers.MemoryBoundedLinkedBlockingQueue;
 import io.airbyte.protocol.models.v0.AirbyteMessage;
 import io.airbyte.protocol.models.v0.StreamDescriptor;
-import java.time.Instant;
-import java.util.*;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
 import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicLong;
-import java.util.concurrent.locks.ReentrantLock;
-import java.util.stream.Stream;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.io.FileUtils;
 
@@ -33,7 +28,7 @@ public class BufferManager {
 
   private final ConcurrentMap<StreamDescriptor, MemoryBoundedLinkedBlockingQueue<AirbyteMessage>> buffers;
   private final BufferManagerEnqueue bufferManagerEnqueue;
-  private final BufferManagerDequeue bufferManagerDequeue;
+  private final BufferDequeue bufferManagerDequeue;
   private final GlobalMemoryManager memoryManager;
   private final ScheduledExecutorService debugLoop = Executors.newSingleThreadScheduledExecutor();
 
@@ -41,7 +36,7 @@ public class BufferManager {
     memoryManager = new GlobalMemoryManager(TOTAL_QUEUES_MAX_SIZE_LIMIT_BYTES);
     buffers = new ConcurrentHashMap<>();
     bufferManagerEnqueue = new BufferManagerEnqueue(memoryManager, buffers);
-    bufferManagerDequeue = new BufferManagerDequeue(memoryManager, buffers);
+    bufferManagerDequeue = new BufferDequeue(memoryManager, buffers);
     debugLoop.scheduleAtFixedRate(this::printQueueInfo, 0, 10, TimeUnit.SECONDS);
   }
 
@@ -49,7 +44,7 @@ public class BufferManager {
     return bufferManagerEnqueue;
   }
 
-  public BufferManagerDequeue getBufferManagerDequeue() {
+  public BufferDequeue getBufferManagerDequeue() {
     return bufferManagerDequeue;
   }
 
@@ -115,91 +110,6 @@ public class BufferManager {
           queue.setMaxMemoryUsage(queue.getMaxMemoryUsage() + freeMem);
         }
         addedToQueue = queue.offer(message, messageSize);
-      }
-    }
-
-  }
-
-  // todo (cgardens) - make all the metadata methods more efficient.
-  static class BufferManagerDequeue {
-
-    private final GlobalMemoryManager memoryManager;
-    private final ConcurrentMap<StreamDescriptor, MemoryBoundedLinkedBlockingQueue<AirbyteMessage>> buffers;
-    private final ConcurrentMap<StreamDescriptor, ReentrantLock> bufferLocks;
-
-    public BufferManagerDequeue(final GlobalMemoryManager memoryManager,
-                                final ConcurrentMap<StreamDescriptor, MemoryBoundedLinkedBlockingQueue<AirbyteMessage>> buffers) {
-      this.memoryManager = memoryManager;
-      this.buffers = buffers;
-      bufferLocks = new ConcurrentHashMap<>();
-    }
-
-    public Map<StreamDescriptor, MemoryBoundedLinkedBlockingQueue<AirbyteMessage>> getBuffers() {
-      return new HashMap<>(buffers);
-    }
-
-    private MemoryBoundedLinkedBlockingQueue<AirbyteMessage> getBuffer(final StreamDescriptor streamDescriptor) {
-      return buffers.get(streamDescriptor);
-    }
-
-    public long getTotalGlobalQueueSizeBytes() {
-      return buffers.values().stream().map(MemoryBoundedLinkedBlockingQueue::getCurrentMemoryUsage).mapToLong(Long::longValue).sum();
-    }
-
-    public long getQueueSizeInRecords(final StreamDescriptor streamDescriptor) {
-      return getBuffer(streamDescriptor).size();
-    }
-
-    public long getQueueSizeBytes(final StreamDescriptor streamDescriptor) {
-      return getBuffer(streamDescriptor).getCurrentMemoryUsage();
-    }
-
-    public Optional<Instant> getTimeOfLastRecord(final StreamDescriptor streamDescriptor) {
-      return getBuffer(streamDescriptor).getTimeOfLastMessage();
-    }
-
-    public Batch take(final StreamDescriptor streamDescriptor, final long bytesToRead) {
-      final var queue = buffers.get(streamDescriptor);
-
-      if (!bufferLocks.containsKey(streamDescriptor)) {
-        bufferLocks.put(streamDescriptor, new ReentrantLock());
-      }
-
-      bufferLocks.get(streamDescriptor).lock();
-      try {
-        final AtomicLong bytesRead = new AtomicLong();
-
-        final var s = Stream.generate(() -> {
-          try {
-            return queue.poll(5, TimeUnit.MILLISECONDS);
-          } catch (final InterruptedException e) {
-            throw new RuntimeException(e);
-          }
-        }).takeWhile(memoryItem -> {
-          // if no new records after waiting, the stream is done.
-          if (memoryItem == null) {
-            return false;
-          }
-
-          // otherwise pull records until we hit the memory limit.
-          final long newSize = memoryItem.size() + bytesRead.get();
-          if (newSize <= bytesToRead) {
-            bytesRead.addAndGet(memoryItem.size());
-            return true;
-          } else {
-            return false;
-          }
-        }).map(MemoryItem::item)
-            .toList()
-            .stream();
-
-        // todo (cgardens) - possible race where in between pulling records and new records going in that we
-        // reset the limit to be lower than number of bytes already in the queue. probably not a big deal.
-        queue.setMaxMemoryUsage(queue.getMaxMemoryUsage() - bytesRead.get());
-
-        return new Batch(s, bytesRead.get(), memoryManager);
-      } finally {
-        bufferLocks.get(streamDescriptor).unlock();
       }
     }
